implementácia api users

+++ src/OPCAIC.ApiService/Exceptions/NotFoundException.cs
@@ -4,7 +4,7 @@ namespace OPCAIC.ApiService.Exceptions
 {
 	public class NotFoundException : ApiException
 	{
-		public NotFoundException(string resourceName, int resourceId) :
+		public NotFoundException(string resourceName, long resourceId) :
 			base(StatusCodes.Status404NotFound,
 				$"Resource '{resourceName}' with id {resourceId} was not found.")
 		{
